fix dist fleet ctr ut
